Add some stubbed libc functions and syscalls
* fix struct dirent to use POSIX-compliant naming
* add st_nlink and st_rdev to struct stat; st_dev is still used
  incorrectly though
* add umount, umount2, fstatvfs, statvfs stubbed syscalls
* add stub for atexit, getopt*, mntent, strtok*, feof
* add (partially) implemented ferror, fread, fwrite, clearerr, strspn,
  memcmp, isprint, strerror
* add alloca.h, assert.h, termios.h and sys/select.h
* make gid_t an int
* add test for strspn

+++ include/kernel/types.h
@@ -12,7 +12,7 @@ typedef uint32_t    clockid_t;
 typedef uint16_t    dev_t;
 typedef uint32_t    fsblkcnt_t;
 typedef uint32_t    fsfilcnt_t;
-typedef uint16_t    gid_t;
+typedef int         gid_t;
 typedef uint32_t    id_t;
 typedef uint32_t    ino_t;
 typedef uint16_t    mode_t;
